Added db checks in shell

+++ LiteDB/Shell/Commands/Files/Delete.cs
@@ -16,6 +16,8 @@ namespace LiteDB.Shell.Commands
 
         public void Execute(LiteEngine db, StringScanner s, Display display)
         {
+            if (db == null) throw new LiteException("No database");
+
             var id = this.ReadId(s);
 
             display.WriteBson(db.FileStorage.Delete(id));
